NejakaMapa13.6.17

+++ NejakaMapa/Models/Souradnice.cs
@@ -8,7 +8,7 @@ namespace NejakaMapa.Models
     public class Souradnice
     {
         public Guid Id { get; set; }
-        public float Lat { get; set; }
-        public float Lng { get; set; }
+        public double Lat { get; set; }
+        public double Lng { get; set; }
     }
 }
